parte do free
Upon receiving this message, a peer that has a local copy of the chunk
shall update its local count of this chunk. If this count drops below
the desired replication degree of that chunk, it shall initiate the
chunk backup subprotocol after a random delay uniformly distributed
between 0 and 400 ms. A peer receiving the PUTCHUNK message, should back
off and restrain from starting yet another backup subprotocol for the
same file chunk.